+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Fixed button label when adding/removing collections/sellers/etc to/from org
+
 ## [1.24.4] - 2023-05-12
 
 ### Fixed